refactor(Panzoid): move enums and shared functions to 'shared.ts'
