+++ websites/P/Panzoid/metadata.json
@@ -12,7 +12,7 @@
 		"app.panzoid.com",
 		"panzoid.com"
 	],
-	"version": "0.3.1",
+	"version": "0.3.2",
 	"logo": "https://i.postimg.cc/3wgKPWjY/panzoid-discord.webp",
 	"thumbnail": "https://i.postimg.cc/3wgKPWjY/panzoid-discord.webp",
 	"color": "#000000",
